[COZY-69] refactor: 사용자 상세정보 등록 Feedback 1차 반영
[COZY-69] refactor: Controller에 @Operation 사용해 Swagger 정보 추가
[COZY-69] refactor: Controller에 TODO 추가
[COZY-69] refactor: Request DTO validation 중, 시간에 관한 제약 0~24에서 0~23으로 수정
[COZY-69] refactor: ENUM 제외에 따른, 타입 및 Validation 변경
[COZY-69] refactor: MemberStat의 Member fetchType을 Lazy로 변경
[COZY-69] refactor: Member 엔티티단에 toEntity 메서드 추가해 빌더를 엔티티로 내리기, Service 코드 변경
[COZY-69] refactor: 인터페이스 삭제, 기존 Service 파일명 변경
[COZY-69] refactor: Custom Exception -> General Exception으로 통일하기
[COZY-69] refactor: 기타 불필요한 import 정리
