+++ src/main/java/com/cozymate/cozymate_server/domain/memberstat/controller/MemberStatController.java
@@ -3,10 +3,10 @@ package com.cozymate.cozymate_server.domain.memberstat.controller;
 import com.cozymate.cozymate_server.domain.memberstat.dto.MemberStatRequestDTO;
 import com.cozymate.cozymate_server.domain.memberstat.service.MemberStatCommandService;
 import com.cozymate.cozymate_server.global.response.ApiResponse;
+import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -22,8 +22,10 @@ public class MemberStatController {
     private final MemberStatCommandService memberStatCommandService;
 
     /**
-     * [POST] 사용자 상세정보 저장
+     * [POST] 사용자 상세정보 등록
+     * TODO: member는 추후 시큐리티 인증 객체에서 받아오는 것으로 변경 예정, path 변경 예정 사항("/memberId" -> "/")
      */
+    @Operation(summary = "[포비] 사용자 상세정보 등록", description = "Path Variable로 memberId, body로 사용자 상세정보를 넣어 사용합니다.")
     @PostMapping("/{memberId}")
     public ApiResponse<Long> createMemberStat(
         @PathVariable("memberId") Long memberId,@Valid @RequestBody MemberStatRequestDTO memberStatRequestDTO) {
