--- src/main/java/com/cozymate/cozymate_server/domain/memberstat/enums/Acceptance.java
@@ -1,15 +1,0 @@
-package com.cozymate.cozymate_server.domain.memberstat.enums;
-
-public enum Acceptance {
-    ACCEPT("합격"),
-    WAITING("대기중"),
-    SPARE_NUM("대기번호"),
-
-    ;
-
-    private String acceptance;
-
-    Acceptance(String acceptance) {
-        this.acceptance = acceptance;
-    }
-}
