--- src/main/java/com/cozymate/cozymate_server/domain/memberstat/enums/SmokingState.java
@@ -1,16 +1,0 @@
-package com.cozymate.cozymate_server.domain.memberstat.enums;
-
-public enum SmokingState {
-    NO("X"),
-    CIGARETTE("연초"),
-    ELECTRONIC("전자담배"),
-    QUITTING("끊는중이에요"),
-
-    ;
-
-    private String state;
-
-    SmokingState(String state){
-        this.state = state;
-    }
-}
