--- src/main/java/com/cozymate/cozymate_server/domain/memberstat/service/MemberStatQueryServiceImpl.java
@@ -1,15 +1,0 @@
-package com.cozymate.cozymate_server.domain.memberstat.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-
-@Slf4j
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-@Service
-public class MemberStatQueryServiceImpl implements MemberStatQueryService {
-
-}
